+++ app/src/main/java/eu/kanade/tachiyomi/data/database/mappers/TrackTypeMapping.kt
@@ -9,38 +9,38 @@ import com.pushtorefresh.storio.sqlite.operations.put.DefaultPutResolver
 import com.pushtorefresh.storio.sqlite.queries.DeleteQuery
 import com.pushtorefresh.storio.sqlite.queries.InsertQuery
 import com.pushtorefresh.storio.sqlite.queries.UpdateQuery
-import eu.kanade.tachiyomi.data.database.models.MangaSync
-import eu.kanade.tachiyomi.data.database.models.MangaSyncImpl
-import eu.kanade.tachiyomi.data.database.tables.MangaSyncTable.COL_ID
-import eu.kanade.tachiyomi.data.database.tables.MangaSyncTable.COL_LAST_CHAPTER_READ
-import eu.kanade.tachiyomi.data.database.tables.MangaSyncTable.COL_MANGA_ID
-import eu.kanade.tachiyomi.data.database.tables.MangaSyncTable.COL_REMOTE_ID
-import eu.kanade.tachiyomi.data.database.tables.MangaSyncTable.COL_SCORE
-import eu.kanade.tachiyomi.data.database.tables.MangaSyncTable.COL_STATUS
-import eu.kanade.tachiyomi.data.database.tables.MangaSyncTable.COL_SYNC_ID
-import eu.kanade.tachiyomi.data.database.tables.MangaSyncTable.COL_TITLE
-import eu.kanade.tachiyomi.data.database.tables.MangaSyncTable.COL_TOTAL_CHAPTERS
-import eu.kanade.tachiyomi.data.database.tables.MangaSyncTable.TABLE
+import eu.kanade.tachiyomi.data.database.models.Track
+import eu.kanade.tachiyomi.data.database.models.TrackImpl
+import eu.kanade.tachiyomi.data.database.tables.TrackTable.COL_ID
+import eu.kanade.tachiyomi.data.database.tables.TrackTable.COL_LAST_CHAPTER_READ
+import eu.kanade.tachiyomi.data.database.tables.TrackTable.COL_MANGA_ID
+import eu.kanade.tachiyomi.data.database.tables.TrackTable.COL_REMOTE_ID
+import eu.kanade.tachiyomi.data.database.tables.TrackTable.COL_SCORE
+import eu.kanade.tachiyomi.data.database.tables.TrackTable.COL_STATUS
+import eu.kanade.tachiyomi.data.database.tables.TrackTable.COL_SYNC_ID
+import eu.kanade.tachiyomi.data.database.tables.TrackTable.COL_TITLE
+import eu.kanade.tachiyomi.data.database.tables.TrackTable.COL_TOTAL_CHAPTERS
+import eu.kanade.tachiyomi.data.database.tables.TrackTable.TABLE
 
-class MangaSyncTypeMapping : SQLiteTypeMapping<MangaSync>(
-        MangaSyncPutResolver(),
-        MangaSyncGetResolver(),
-        MangaSyncDeleteResolver()
+class TrackTypeMapping : SQLiteTypeMapping<Track>(
+        TrackPutResolver(),
+        TrackGetResolver(),
+        TrackDeleteResolver()
 )
 
-class MangaSyncPutResolver : DefaultPutResolver<MangaSync>() {
+class TrackPutResolver : DefaultPutResolver<Track>() {
 
-    override fun mapToInsertQuery(obj: MangaSync) = InsertQuery.builder()
+    override fun mapToInsertQuery(obj: Track) = InsertQuery.builder()
             .table(TABLE)
             .build()
 
-    override fun mapToUpdateQuery(obj: MangaSync) = UpdateQuery.builder()
+    override fun mapToUpdateQuery(obj: Track) = UpdateQuery.builder()
             .table(TABLE)
             .where("$COL_ID = ?")
             .whereArgs(obj.id)
             .build()
 
-    override fun mapToContentValues(obj: MangaSync) = ContentValues(9).apply {
+    override fun mapToContentValues(obj: Track) = ContentValues(9).apply {
         put(COL_ID, obj.id)
         put(COL_MANGA_ID, obj.manga_id)
         put(COL_SYNC_ID, obj.sync_id)
@@ -53,9 +53,9 @@ class MangaSyncPutResolver : DefaultPutResolver<MangaSync>() {
     }
 }
 
-class MangaSyncGetResolver : DefaultGetResolver<MangaSync>() {
+class TrackGetResolver : DefaultGetResolver<Track>() {
 
-    override fun mapFromCursor(cursor: Cursor): MangaSync = MangaSyncImpl().apply {
+    override fun mapFromCursor(cursor: Cursor): Track = TrackImpl().apply {
         id = cursor.getLong(cursor.getColumnIndex(COL_ID))
         manga_id = cursor.getLong(cursor.getColumnIndex(COL_MANGA_ID))
         sync_id = cursor.getInt(cursor.getColumnIndex(COL_SYNC_ID))
@@ -68,9 +68,9 @@ class MangaSyncGetResolver : DefaultGetResolver<MangaSync>() {
     }
 }
 
-class MangaSyncDeleteResolver : DefaultDeleteResolver<MangaSync>() {
+class TrackDeleteResolver : DefaultDeleteResolver<Track>() {
 
-    override fun mapToDeleteQuery(obj: MangaSync) = DeleteQuery.builder()
+    override fun mapToDeleteQuery(obj: Track) = DeleteQuery.builder()
             .table(TABLE)
             .where("$COL_ID = ?")
             .whereArgs(obj.id)
