--- app/src/main/java/eu/kanade/tachiyomi/data/mangasync/anilist/model/ALUserLists.kt
@@ -1,6 +1,0 @@
-package eu.kanade.tachiyomi.data.mangasync.anilist.model
-
-data class ALUserLists(val lists: Map<String, List<ALUserManga>>) {
-
-    fun flatten() = lists.values.flatten()
-}
