+++ app/src/main/java/eu/kanade/tachiyomi/data/preference/PreferencesHelper.kt
@@ -7,7 +7,7 @@ import android.preference.PreferenceManager
 import com.f2prateek.rx.preferences.Preference
 import com.f2prateek.rx.preferences.RxSharedPreferences
 import eu.kanade.tachiyomi.R
-import eu.kanade.tachiyomi.data.mangasync.MangaSyncService
+import eu.kanade.tachiyomi.data.track.TrackService
 import eu.kanade.tachiyomi.data.source.Source
 import java.io.File
 
@@ -95,11 +95,11 @@ class PreferencesHelper(context: Context) {
                 .apply()
     }
 
-    fun mangaSyncUsername(sync: MangaSyncService) = prefs.getString(keys.syncUsername(sync.id), "")
+    fun mangaSyncUsername(sync: TrackService) = prefs.getString(keys.syncUsername(sync.id), "")
 
-    fun mangaSyncPassword(sync: MangaSyncService) = prefs.getString(keys.syncPassword(sync.id), "")
+    fun mangaSyncPassword(sync: TrackService) = prefs.getString(keys.syncPassword(sync.id), "")
 
-    fun setMangaSyncCredentials(sync: MangaSyncService, username: String, password: String) {
+    fun setMangaSyncCredentials(sync: TrackService, username: String, password: String) {
         prefs.edit()
                 .putString(keys.syncUsername(sync.id), username)
                 .putString(keys.syncPassword(sync.id), password)
