+++ app/src/main/java/eu/kanade/tachiyomi/data/track/TrackManager.kt
@@ -1,0 +1,24 @@
+package eu.kanade.tachiyomi.data.track
+
+import android.content.Context
+import eu.kanade.tachiyomi.data.track.anilist.Anilist
+import eu.kanade.tachiyomi.data.track.myanimelist.MyAnimeList
+
+class TrackManager(private val context: Context) {
+
+    companion object {
+        const val MYANIMELIST = 1
+        const val ANILIST = 2
+    }
+
+    val myAnimeList = MyAnimeList(context, MYANIMELIST)
+
+    val aniList = Anilist(context, ANILIST)
+
+    val services = listOf(myAnimeList, aniList)
+
+    fun getService(id: Int) = services.find { it.id == id }
+
+    fun hasLoggedServices() = services.any { it.isLogged }
+
+}
