+++ app/src/main/java/eu/kanade/tachiyomi/data/track/anilist/model/ALManga.kt
@@ -1,0 +1,17 @@
+package eu.kanade.tachiyomi.data.track.anilist.model
+
+import eu.kanade.tachiyomi.data.database.models.Track
+import eu.kanade.tachiyomi.data.track.TrackManager
+
+data class ALManga(
+        val id: Int,
+        val title_romaji: String,
+        val type: String,
+        val total_chapters: Int) {
+
+    fun toMangaSync() = Track.create(TrackManager.ANILIST).apply {
+        remote_id = this@ALManga.id
+        title = title_romaji
+        total_chapters = this@ALManga.total_chapters
+    }
+}
