+++ app/src/main/java/eu/kanade/tachiyomi/data/track/anilist/model/ALUserLists.kt
@@ -1,0 +1,6 @@
+package eu.kanade.tachiyomi.data.track.anilist.model
+
+data class ALUserLists(val lists: Map<String, List<ALUserManga>>) {
+
+    fun flatten() = lists.values.flatten()
+}
