+++ app/src/main/java/eu/kanade/tachiyomi/data/track/anilist/model/OAuth.kt
@@ -1,0 +1,11 @@
+package eu.kanade.tachiyomi.data.track.anilist.model
+
+data class OAuth(
+        val access_token: String,
+        val token_type: String,
+        val expires: Long,
+        val expires_in: Long,
+        val refresh_token: String?) {
+
+    fun isExpired() = System.currentTimeMillis() > expires
+}
