+++ app/src/main/java/eu/kanade/tachiyomi/ui/setting/SettingsFragment.kt
@@ -28,7 +28,7 @@ open class SettingsFragment : XpPreferenceFragment() {
         addPreferencesFromResource(R.xml.pref_reader)
         addPreferencesFromResource(R.xml.pref_downloads)
         addPreferencesFromResource(R.xml.pref_sources)
-        addPreferencesFromResource(R.xml.pref_sync)
+        addPreferencesFromResource(R.xml.pref_tracking)
         addPreferencesFromResource(R.xml.pref_advanced)
         addPreferencesFromResource(R.xml.pref_about)
